LONDispatch code complete.  Still need to debug whenever ready.
git-svn-id: https://cotprojects.marine.usf.edu/svn/MiniSEAS/Gumstix/Source@64 63cbca5c-2ead-4e1b-9715-33247cc11c71

+++ StixServer/MethodParser/MethodNodesTable.c
@@ -17,7 +17,7 @@ s_node* current = NULL;
 #define MAXCONTROLFLOWDEPTH 20
 stack* openControlStack = NULL;
 
-s_node* buildCommandNode(int argc, void* argv, void (*command)(int,void*))
+s_node* buildCommandNode(unsigned long argc, void* argv, void (*command)(unsigned long,void*))
 {
     s_node* node = malloc(sizeof(s_node));
     node->argc = argc;
